% File MedOr/man/schizophrenia.Rd
% Part of the MedOr package for R (http://www.R-project.org)
% Copyright (C) 2012, Adriano Polpo, Carlos A. de B. Pereira.
% Distributed under GPL 3 or later

\name{schizophrenia}

%%%
\alias{schizophrenia}
\title{Data set Schizophrenia}
\description{
Data extracted from study Altar A of the Stanley Medical Research Institute online genomics database (SMRIDB).
}
\value{
  The data variables are:
  \item{           group}{1: for control group, and 2: for schizophrenic.}
  \item{others variables}{each column represents the measure of a microarray probe.}
}
\source{
  The Stanley Medical Research Institute online genomics database, \url{www.stanleygenomics.org}.
}
\references{
  Higgs, B.W., Elashoff, M., Richman, S. and Barci, B. (2006)
  \emph{An online database for brain disease research}.
  BMC Genomics 2006, 7:70,
  (DOI \href{http://dx.doi.org/10.1186/1471-2164-7-70}{10.1186/1471-2164-7-70}).
}
